pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "labelscope"
version = "0.0.1"
description = "Python SDK and viewer to preview thermal printer labels (EPL2, with ZPL planned) from their source code."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [
    { name = "Jhonatan", email = "it@classicalpaca.com" },
]
keywords = ["epl2", "zpl", "thermal-printer", "zebra", "label", "preview", "barcode"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Printing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "Pillow>=10.0",
]

[project.optional-dependencies]
barcodes = [
    "treepoem>=3.24",
    "python-barcode",
]
dev = [
    "pytest",
    "pytest-regressions",
    "ruff",
    "mypy",
]

[project.scripts]
labelscope = "labelscope.cli:main"

[project.urls]
Homepage = "https://github.com/xpertik/labelscope"
Repository = "https://github.com/xpertik/labelscope"
Issues = "https://github.com/xpertik/labelscope/issues"

[tool.hatch.build.targets.wheel]
packages = ["labelscope"]

[tool.hatch.build.targets.wheel.force-include]
"labelscope/core/_fonts/DejaVuSansMono.ttf" = "labelscope/core/_fonts/DejaVuSansMono.ttf"
"labelscope/core/_fonts/DejaVuSansMono-LICENSE.txt" = "labelscope/core/_fonts/DejaVuSansMono-LICENSE.txt"
"labelscope/py.typed" = "labelscope/py.typed"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
